+++ packages/next/src/server/app-render/action-handler.ts
@@ -37,6 +37,7 @@ import {
   NEXT_CACHE_REVALIDATED_TAGS_HEADER,
   NEXT_CACHE_REVALIDATE_TAG_TOKEN_HEADER,
 } from '../../lib/constants'
+import type { AppRenderContext, GenerateFlight } from './app-render'
 
 function nodeToWebReadableStream(nodeReadable: import('stream').Readable) {
   if (process.env.NEXT_RUNTIME !== 'edge') {
@@ -246,21 +247,18 @@ export async function handleAction({
   staticGenerationStore,
   requestStore,
   serverActionsBodySizeLimit,
+  ctx,
 }: {
   req: IncomingMessage
   res: ServerResponse
   ComponentMod: any
   page: string
   serverActionsManifest: any
-  generateFlight: (options: {
-    actionResult: ActionResult
-    formState?: any
-    skipFlight: boolean
-    asNotFound?: boolean
-  }) => Promise<RenderResult>
+  generateFlight: GenerateFlight
   staticGenerationStore: StaticGenerationStore
   requestStore: RequestStore
   serverActionsBodySizeLimit?: SizeLimit
+  ctx: AppRenderContext
 }): Promise<
   | undefined
   | {
@@ -449,7 +447,7 @@ export async function handleAction({
             requestStore,
           })
 
-          actionResult = await generateFlight({
+          actionResult = await generateFlight(ctx, {
             actionResult: Promise.resolve(returnVal),
             // if the page was not revalidated, we can skip the rendering the flight tree
             skipFlight: !staticGenerationStore.pathWasRevalidated,
